Add files via upload
This project will be analyzing apps that are free and get most of their revenue through in app ads. The objective is to attract more users to increase revenue for a company I'm working with. To do that, we will need to see what type of free apps perform well on both the IOS App Store and the Google Play Store.